#138 検索Input部分をコンポーネント化
https://github.com/ysds/acnh-gachi-complete/pull/138#issuecomment-994774349
